Separate overview from playing view.
Make the playing view show one section at a time, large and in the middle.

The CSS wrangling needed to make this happen makes me regret doing all this
with stock HTML. I should have just done it in Canvas with Pixi.js or
something. It's already a messy disaster. I hate the web.