Adding Pattern Trainer.
This is not a drill, but rather generates random 1e+a patterns, plays them for
you to listen, and allows you to repeat them. The patterns can be easy (repeat
a single measure four times) medium (two unique measures, repeated) or hard
(four unique measures). The patterns can be shown or not shown at different
times to add challenge as well.